+++ src-docs/src/views/suggest/suggest.js
@@ -1,21 +1,15 @@
 import React, { Component } from 'react';
 
 import {
-  EuiButton,
   EuiButtonEmpty,
   EuiFlexGroup,
   EuiFlexItem,
-  EuiIcon,
-  EuiListGroup,
-  EuiListGroupItem,
-  EuiPopover,
-  EuiPopoverFooter,
-  EuiPopoverTitle,
+  EuiRadioGroup,
   EuiSuggest,
-  EuiSelect,
   EuiSpacer,
-  EuiText,
 } from '../../../../src/components';
+
+import makeId from '../../../../src/components/form/form_row/make_id';
 
 const shortDescription = 'This is the description';
 
@@ -54,25 +48,28 @@ export default class extends Component {
   constructor(props) {
     super(props);
 
-    this.options = [
-      { value: 'notYetSaved', text: 'Not yet saved' },
-      { value: 'saved', text: 'Saved' },
-      { value: 'noNewChanges', text: 'No new changes' },
-      { value: 'isLoading', text: 'Loading' },
+    const idPrefix = makeId();
+
+    this.radios = [
+      { id: `${idPrefix}0`, value: 'notYetSaved', label: 'Not yet saved' },
+      { id: `${idPrefix}1`, value: 'saved', label: 'Saved' },
+      { id: `${idPrefix}2`, value: 'noNewChanges', label: 'No new changes' },
+      { id: `${idPrefix}3`, value: 'isLoading', label: 'Loading' },
     ];
 
     this.state = {
-      isHashtagPopoverOpen: false,
       status: 'notYetSaved',
+      radioIdSelected: `${idPrefix}0`,
       value: '',
     };
   }
 
-  closeHashtagPopover() {
+  onChange = optionId => {
     this.setState({
-      isHashtagPopoverOpen: false,
+      radioIdSelected: optionId,
+      status: this.radios.find(x => x.id === optionId).value,
     });
-  }
+  };
 
   onSelectChange(e) {
     this.setState({
@@ -86,86 +83,17 @@ export default class extends Component {
     });
   }
 
-  onButtonClick() {
-    this.setState({
-      isHashtagPopoverOpen: !this.state.isHashtagPopoverOpen,
-    });
-  }
-
   render() {
-    const hashtagButton = (
-      <EuiButtonEmpty
-        onClick={this.onButtonClick.bind(this)}
-        iconType="arrowDown"
-        iconSide="right">
-        <EuiIcon type="number" />
-      </EuiButtonEmpty>
-    );
-
-    const hashtag = (
-      <EuiPopover
-        id="popover"
-        button={hashtagButton}
-        isOpen={this.state.isHashtagPopoverOpen}
-        anchorPosition="downLeft"
-        panelPaddingSize="none"
-        closePopover={this.closeHashtagPopover.bind(this)}>
-        <EuiPopoverTitle>SAVED QUERIES</EuiPopoverTitle>
-        <div>
-          <EuiText
-            size="s"
-            color="subdued"
-            className="savedQueryManagement__text">
-            <p>Save query text and filters that you want to use again.</p>
-          </EuiText>
-          <div className="savedQueryManagement__listWrapper">
-            <EuiListGroup className="savedQueryManagement__list" flush={true}>
-              <EuiListGroupItem
-                extraAction={{
-                  color: 'danger',
-                  iconType: 'trash',
-                  iconSize: 's',
-                }}
-                href="#"
-                label="Popular shoes in America"
-              />
-              <EuiListGroupItem
-                extraAction={{
-                  color: 'danger',
-                  iconType: 'trash',
-                  iconSize: 's',
-                }}
-                href="#"
-                label="Popular shirts in Canada"
-              />
-            </EuiListGroup>
-          </div>
-          {this.state.value !== '' ? (
-            <EuiPopoverFooter>
-              <EuiFlexGroup direction="rowReverse" alignItems="center">
-                <EuiFlexItem grow={false}>
-                  <EuiButton fill>Save</EuiButton>
-                </EuiFlexItem>
-              </EuiFlexGroup>
-            </EuiPopoverFooter>
-          ) : (
-            undefined
-          )}
-        </div>
-      </EuiPopover>
-    );
-
     const append = <EuiButtonEmpty>KQL</EuiButtonEmpty>;
 
     return (
       <div>
         <EuiFlexGroup>
           <EuiFlexItem>
-            <EuiSelect
-              options={this.options}
-              value={this.state.status}
-              onChange={this.onSelectChange.bind(this)}
-              aria-label="Use aria labels when no actual label is in use"
+            <EuiRadioGroup
+              options={this.radios}
+              idSelected={this.state.radioIdSelected}
+              onChange={this.onChange}
             />
           </EuiFlexItem>
         </EuiFlexGroup>
@@ -174,7 +102,6 @@ export default class extends Component {
           <EuiFlexItem>
             <EuiSuggest
               status={this.state.status}
-              prefix={hashtag}
               sendInputValue={this.getInputValue.bind(this)}
               append={append}
               suggestions={sampleItems}